+++ src/handlers/ReactNative106.ts
@@ -3,7 +3,9 @@ import { Logger } from '../Logger';
 import * as utils from '../utils';
 import * as ortc from '../ortc';
 import * as sdpCommonUtils from './sdp/commonUtils';
-import * as sdpPlanBUtils from './sdp/planBUtils';
+import * as sdpUnifiedPlanUtils from './sdp/unifiedPlanUtils';
+import * as ortcUtils from './ortc/utils';
+import { InvalidStateError } from '../errors';
 import {
 	type HandlerFactory,
 	HandlerInterface,
@@ -18,16 +20,23 @@ import {
 	type HandlerReceiveDataChannelResult,
 } from './HandlerInterface';
 import { RemoteSdp } from './sdp/RemoteSdp';
+import { parse as parseScalabilityMode } from '../scalabilityModes';
 import type { IceParameters, DtlsRole } from '../Transport';
-import type { RtpCapabilities, RtpParameters } from '../RtpParameters';
+import type {
+	RtpCapabilities,
+	RtpParameters,
+	RtpEncodingParameters,
+} from '../RtpParameters';
 import type { SctpCapabilities, SctpStreamParameters } from '../SctpParameters';
 
-const logger = new Logger('Chrome67');
+const logger = new Logger('ReactNative106');
 
-const NAME = 'Chrome67';
+const NAME = 'ReactNative106';
 const SCTP_NUM_STREAMS = { OS: 1024, MIS: 1024 };
 
-export class Chrome67 extends HandlerInterface {
+export class ReactNative106 extends HandlerInterface {
+	// Closed flag.
+	private _closed = false;
 	// Handler direction.
 	private _direction?: 'send' | 'recv';
 	// Remote SDP handler.
@@ -42,23 +51,11 @@ export class Chrome67 extends HandlerInterface {
 	private _forcedLocalDtlsRole?: DtlsRole;
 	// RTCPeerConnection instance.
 	private _pc: any;
+	// Map of RTCTransceivers indexed by MID.
+	private readonly _mapMidTransceiver: Map<string, RTCRtpTransceiver> =
+		new Map();
 	// Local stream for sending.
 	private readonly _sendStream = new MediaStream();
-	// Map of RTCRtpSender indexed by localId.
-	private readonly _mapSendLocalIdRtpSender: Map<string, RTCRtpSender> =
-		new Map();
-	// Next sending localId.
-	private _nextSendLocalId = 0;
-	// Map of MID, RTP parameters and RTCRtpReceiver indexed by local id.
-	// Value is an Object with mid, rtpParameters and rtpReceiver.
-	private readonly _mapRecvLocalIdInfo: Map<
-		string,
-		{
-			mid: string;
-			rtpParameters: RtpParameters;
-			rtpReceiver: RTCRtpReceiver;
-		}
-	> = new Map();
 	// Whether a DataChannel m=application section has been created.
 	private _hasDataChannelMediaSection = false;
 	// Sending DataChannel id value counter. Incremented for each new DataChannel.
@@ -70,7 +67,7 @@ export class Chrome67 extends HandlerInterface {
 	 * Creates a factory function.
 	 */
 	static createFactory(): HandlerFactory {
-		return (): Chrome67 => new Chrome67();
+		return (): ReactNative106 => new ReactNative106();
 	}
 
 	constructor() {
@@ -83,6 +80,17 @@ export class Chrome67 extends HandlerInterface {
 
 	close(): void {
 		logger.debug('close()');
+
+		if (this._closed) {
+			return;
+		}
+
+		this._closed = true;
+
+		// Free/dispose native MediaStream but DO NOT free/dispose native
+		// MediaStreamTracks (that is parent's business).
+		// @ts-expect-error --- Proprietary API in react-native-webrtc.
+		this._sendStream.release(/* releaseTracks */ false);
 
 		// Close RTCPeerConnection.
 		if (this._pc) {
@@ -102,14 +110,14 @@ export class Chrome67 extends HandlerInterface {
 			iceTransportPolicy: 'all',
 			bundlePolicy: 'max-bundle',
 			rtcpMuxPolicy: 'require',
-			sdpSemantics: 'plan-b',
+			sdpSemantics: 'unified-plan',
 		});
 
 		try {
-			const offer = await pc.createOffer({
-				offerToReceiveAudio: true,
-				offerToReceiveVideo: true,
-			});
+			pc.addTransceiver('audio');
+			pc.addTransceiver('video');
+
+			const offer = await pc.createOffer();
 
 			try {
 				pc.close();
@@ -119,6 +127,9 @@ export class Chrome67 extends HandlerInterface {
 			const nativeRtpCapabilities = sdpCommonUtils.extractRtpCapabilities({
 				sdpObject,
 			});
+
+			// libwebrtc supports NACK for OPUS but doesn't announce it.
+			ortcUtils.addNackSupportForOpus(nativeRtpCapabilities);
 
 			return nativeRtpCapabilities;
 		} catch (error) {
@@ -150,6 +161,8 @@ export class Chrome67 extends HandlerInterface {
 		proprietaryConstraints,
 		extendedRtpCapabilities,
 	}: HandlerRunOptions): void {
+		this.assertNotClosed();
+
 		logger.debug('run()');
 
 		this._direction = direction;
@@ -159,7 +172,6 @@ export class Chrome67 extends HandlerInterface {
 			iceCandidates,
 			dtlsParameters,
 			sctpParameters,
-			planB: true,
 		});
 
 		this._sendingRtpParametersByKind = {
@@ -189,7 +201,7 @@ export class Chrome67 extends HandlerInterface {
 				iceTransportPolicy: iceTransportPolicy ?? 'all',
 				bundlePolicy: 'max-bundle',
 				rtcpMuxPolicy: 'require',
-				sdpSemantics: 'plan-b',
+				sdpSemantics: 'unified-plan',
 				...additionalSettings,
 			},
 			proprietaryConstraints
@@ -253,6 +265,8 @@ export class Chrome67 extends HandlerInterface {
 	}
 
 	async updateIceServers(iceServers: RTCIceServer[]): Promise<void> {
+		this.assertNotClosed();
+
 		logger.debug('updateIceServers()');
 
 		const configuration = this._pc.getConfiguration();
@@ -263,6 +277,8 @@ export class Chrome67 extends HandlerInterface {
 	}
 
 	async restartIce(iceParameters: IceParameters): Promise<void> {
+		this.assertNotClosed();
+
 		logger.debug('restartIce()');
 
 		// Provide the remote SDP handler with new remote ICE parameters.
@@ -312,6 +328,8 @@ export class Chrome67 extends HandlerInterface {
 	}
 
 	async getTransportStats(): Promise<RTCStatsReport> {
+		this.assertNotClosed();
+
 		return this._pc.getStats();
 	}
 
@@ -320,20 +338,49 @@ export class Chrome67 extends HandlerInterface {
 		encodings,
 		codecOptions,
 		codec,
+		onRtpSender,
 	}: HandlerSendOptions): Promise<HandlerSendResult> {
+		this.assertNotClosed();
 		this.assertSendDirection();
 
 		logger.debug('send() [kind:%s, track.id:%s]', track.kind, track.id);
 
-		if (codec) {
-			logger.warn(
-				'send() | codec selection is not available in %s handler',
-				this.name
-			);
+		if (encodings && encodings.length > 1) {
+			encodings.forEach((encoding: RtpEncodingParameters, idx: number) => {
+				encoding.rid = `r${idx}`;
+			});
 		}
 
-		this._sendStream.addTrack(track);
-		this._pc.addTrack(track, this._sendStream);
+		const sendingRtpParameters = utils.clone<RtpParameters>(
+			this._sendingRtpParametersByKind![track.kind]!
+		);
+
+		// This may throw.
+		sendingRtpParameters.codecs = ortc.reduceCodecs(
+			sendingRtpParameters.codecs,
+			codec
+		);
+
+		const sendingRemoteRtpParameters = utils.clone<RtpParameters>(
+			this._sendingRemoteRtpParametersByKind![track.kind]!
+		);
+
+		// This may throw.
+		sendingRemoteRtpParameters.codecs = ortc.reduceCodecs(
+			sendingRemoteRtpParameters.codecs,
+			codec
+		);
+
+		const mediaSectionIdx = this._remoteSdp!.getNextMediaSectionIdx();
+		const transceiver = this._pc.addTransceiver(track, {
+			direction: 'sendonly',
+			streams: [this._sendStream],
+			sendEncodings: encodings,
+		});
+
+		if (onRtpSender) {
+			onRtpSender(transceiver.sender);
+		}
 
 		let offer = await this._pc.createOffer();
 		let localSdpObject = sdpTransform.parse(offer.sdp);
@@ -345,21 +392,6 @@ export class Chrome67 extends HandlerInterface {
 		}
 
 		let offerMediaObject;
-		const sendingRtpParameters = utils.clone<RtpParameters>(
-			this._sendingRtpParametersByKind![track.kind]!
-		);
-
-		sendingRtpParameters.codecs = ortc.reduceCodecs(
-			sendingRtpParameters.codecs
-		);
-
-		const sendingRemoteRtpParameters = utils.clone<RtpParameters>(
-			this._sendingRemoteRtpParametersByKind![track.kind]!
-		);
-
-		sendingRemoteRtpParameters.codecs = ortc.reduceCodecs(
-			sendingRemoteRtpParameters.codecs
-		);
 
 		if (!this._transportReady) {
 			await this.setupTransport({
@@ -368,18 +400,28 @@ export class Chrome67 extends HandlerInterface {
 			});
 		}
 
-		if (track.kind === 'video' && encodings && encodings.length > 1) {
-			logger.debug('send() | enabling simulcast');
+		// Special case for VP9 with SVC.
+		let hackVp9Svc = false;
 
+		const layers = parseScalabilityMode(
+			(encodings ?? [{}])[0]!.scalabilityMode
+		);
+
+		if (
+			encodings &&
+			encodings.length === 1 &&
+			layers.spatialLayers > 1 &&
+			sendingRtpParameters.codecs[0]!.mimeType.toLowerCase() === 'video/vp9'
+		) {
+			logger.debug('send() | enabling legacy simulcast for VP9 SVC');
+
+			hackVp9Svc = true;
 			localSdpObject = sdpTransform.parse(offer.sdp);
-			offerMediaObject = localSdpObject.media.find(
-				(m: any) => m.type === 'video'
-			);
+			offerMediaObject = localSdpObject.media[mediaSectionIdx.idx];
 
-			sdpPlanBUtils.addLegacySimulcast({
+			sdpUnifiedPlanUtils.addLegacySimulcast({
 				offerMediaObject,
-				track,
-				numStreams: encodings.length,
+				numStreams: layers.spatialLayers,
 			});
 
 			offer = { type: 'offer', sdp: sdpTransform.write(localSdpObject) };
@@ -389,44 +431,79 @@ export class Chrome67 extends HandlerInterface {
 
 		await this._pc.setLocalDescription(offer);
 
+		// We can now get the transceiver.mid.
+		// NOTE: We cannot read generated MID on iOS react-native-webrtc 111.0.0
+		// because transceiver.mid is not available until setRemoteDescription()
+		// is called, so this is best effort.
+		// Issue: https://github.com/react-native-webrtc/react-native-webrtc/issues/1404
+		// NOTE: So let's fill MID in sendingRtpParameters later.
+		// NOTE: This is fixed in react-native-webrtc 111.0.3.
+		let localId = transceiver.mid ?? undefined;
+
+		if (!localId) {
+			logger.warn(
+				'send() | missing transceiver.mid (bug in react-native-webrtc, using a workaround'
+			);
+		}
+
+		// Set MID.
+		// NOTE: As per above, it could be unset yet.
+		sendingRtpParameters.mid = localId;
+
 		localSdpObject = sdpTransform.parse(this._pc.localDescription.sdp);
-		offerMediaObject = localSdpObject.media.find(
-			(m: any) => m.type === track.kind
-		);
+		offerMediaObject = localSdpObject.media[mediaSectionIdx.idx];
 
 		// Set RTCP CNAME.
 		sendingRtpParameters.rtcp!.cname = sdpCommonUtils.getCname({
 			offerMediaObject,
 		});
 
-		// Set RTP encodings.
-		sendingRtpParameters.encodings = sdpPlanBUtils.getRtpEncodings({
-			offerMediaObject,
-			track,
-		});
+		// Set RTP encodings by parsing the SDP offer if no encodings are given.
+		if (!encodings) {
+			sendingRtpParameters.encodings = sdpUnifiedPlanUtils.getRtpEncodings({
+				offerMediaObject,
+			});
+		}
+		// Set RTP encodings by parsing the SDP offer and complete them with given
+		// one if just a single encoding has been given.
+		else if (encodings.length === 1) {
+			let newEncodings = sdpUnifiedPlanUtils.getRtpEncodings({
+				offerMediaObject,
+			});
 
-		// Complete encodings with given values.
-		if (encodings) {
-			for (let idx = 0; idx < sendingRtpParameters.encodings.length; ++idx) {
-				if (encodings[idx]) {
-					Object.assign(sendingRtpParameters.encodings[idx]!, encodings[idx]);
-				}
+			Object.assign(newEncodings[0]!, encodings[0]);
+
+			// Hack for VP9 SVC.
+			if (hackVp9Svc) {
+				newEncodings = [newEncodings[0]!];
 			}
+
+			sendingRtpParameters.encodings = newEncodings;
+		}
+		// Otherwise if more than 1 encoding are given use them verbatim.
+		else {
+			sendingRtpParameters.encodings = encodings;
 		}
 
-		// If VP8 and there is effective simulcast, add scalabilityMode to each
-		// encoding.
+		// If VP8 or H264 and there is effective simulcast, add scalabilityMode to
+		// each encoding.
 		if (
 			sendingRtpParameters.encodings.length > 1 &&
-			sendingRtpParameters.codecs[0]!.mimeType.toLowerCase() === 'video/vp8'
+			(sendingRtpParameters.codecs[0]!.mimeType.toLowerCase() === 'video/vp8' ||
+				sendingRtpParameters.codecs[0]!.mimeType.toLowerCase() === 'video/h264')
 		) {
 			for (const encoding of sendingRtpParameters.encodings) {
-				encoding.scalabilityMode = 'L1T3';
+				if (encoding.scalabilityMode) {
+					encoding.scalabilityMode = `L1T${layers.temporalLayers}`;
+				} else {
+					encoding.scalabilityMode = 'L1T3';
+				}
 			}
 		}
 
 		this._remoteSdp!.send({
 			offerMediaObject,
+			reuseMid: mediaSectionIdx.reuseMid,
 			offerRtpParameters: sendingRtpParameters,
 			answerRtpParameters: sendingRemoteRtpParameters,
 			codecOptions,
@@ -441,42 +518,53 @@ export class Chrome67 extends HandlerInterface {
 
 		await this._pc.setRemoteDescription(answer);
 
-		const localId = String(this._nextSendLocalId);
+		// Follow up of iOS react-native-webrtc 111.0.0 issue told above. Now yes,
+		// we can read generated MID (if not done above) and fill sendingRtpParameters.
+		// NOTE: This is fixed in react-native-webrtc 111.0.3 so this block isn't
+		// needed starting from that version.
+		if (!localId) {
+			localId = transceiver.mid;
+			sendingRtpParameters.mid = localId;
+		}
 
-		this._nextSendLocalId++;
-
-		const rtpSender = this._pc
-			.getSenders()
-			.find((s: RTCRtpSender) => s.track === track);
-
-		// Insert into the map.
-		this._mapSendLocalIdRtpSender.set(localId, rtpSender);
+		// Store in the map.
+		this._mapMidTransceiver.set(localId, transceiver);
 
 		return {
-			localId: localId,
+			localId,
 			rtpParameters: sendingRtpParameters,
-			rtpSender,
+			rtpSender: transceiver.sender,
 		};
 	}
 
 	async stopSending(localId: string): Promise<void> {
 		this.assertSendDirection();
 
+		if (this._closed) {
+			return;
+		}
+
 		logger.debug('stopSending() [localId:%s]', localId);
 
-		const rtpSender = this._mapSendLocalIdRtpSender.get(localId);
+		const transceiver = this._mapMidTransceiver.get(localId);
 
-		if (!rtpSender) {
-			throw new Error('associated RTCRtpSender not found');
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
 		}
 
-		this._pc.removeTrack(rtpSender);
+		void transceiver.sender.replaceTrack(null);
 
-		if (rtpSender.track) {
-			this._sendStream.removeTrack(rtpSender.track);
+		this._pc.removeTrack(transceiver.sender);
+
+		const mediaSectionClosed = this._remoteSdp!.closeMediaSection(
+			transceiver.mid!
+		);
+
+		if (mediaSectionClosed) {
+			try {
+				transceiver.stop();
+			} catch (error) {}
 		}
-
-		this._mapSendLocalIdRtpSender.delete(localId);
 
 		const offer = await this._pc.createOffer();
 
@@ -485,26 +573,7 @@ export class Chrome67 extends HandlerInterface {
 			offer
 		);
 
-		try {
-			await this._pc.setLocalDescription(offer);
-		} catch (error) {
-			// NOTE: If there are no sending tracks, setLocalDescription() will fail with
-			// "Failed to create channels". If so, ignore it.
-			if (this._sendStream.getTracks().length === 0) {
-				logger.warn(
-					'stopSending() | ignoring expected error due no sending tracks: %s',
-					(error as Error).toString()
-				);
-
-				return;
-			}
-
-			throw error;
-		}
-
-		if (this._pc.signalingState === 'stable') {
-			return;
-		}
+		await this._pc.setLocalDescription(offer);
 
 		const answer = { type: 'answer', sdp: this._remoteSdp!.getSdp() };
 
@@ -514,22 +583,84 @@ export class Chrome67 extends HandlerInterface {
 		);
 
 		await this._pc.setRemoteDescription(answer);
+
+		this._mapMidTransceiver.delete(localId);
 	}
 
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
 	async pauseSending(localId: string): Promise<void> {
-		// Unimplemented.
+		this.assertNotClosed();
+		this.assertSendDirection();
+
+		logger.debug('pauseSending() [localId:%s]', localId);
+
+		const transceiver = this._mapMidTransceiver.get(localId);
+
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
+		}
+
+		transceiver.direction = 'inactive';
+		this._remoteSdp!.pauseMediaSection(localId);
+
+		const offer = await this._pc.createOffer();
+
+		logger.debug(
+			'pauseSending() | calling pc.setLocalDescription() [offer:%o]',
+			offer
+		);
+
+		await this._pc.setLocalDescription(offer);
+
+		const answer = { type: 'answer', sdp: this._remoteSdp!.getSdp() };
+
+		logger.debug(
+			'pauseSending() | calling pc.setRemoteDescription() [answer:%o]',
+			answer
+		);
+
+		await this._pc.setRemoteDescription(answer);
 	}
 
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
 	async resumeSending(localId: string): Promise<void> {
-		// Unimplemented.
+		this.assertNotClosed();
+		this.assertSendDirection();
+
+		logger.debug('resumeSending() [localId:%s]', localId);
+
+		const transceiver = this._mapMidTransceiver.get(localId);
+
+		this._remoteSdp!.resumeSendingMediaSection(localId);
+
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
+		}
+
+		transceiver.direction = 'sendonly';
+
+		const offer = await this._pc.createOffer();
+
+		logger.debug(
+			'resumeSending() | calling pc.setLocalDescription() [offer:%o]',
+			offer
+		);
+
+		await this._pc.setLocalDescription(offer);
+
+		const answer = { type: 'answer', sdp: this._remoteSdp!.getSdp() };
+
+		logger.debug(
+			'resumeSending() | calling pc.setRemoteDescription() [answer:%o]',
+			answer
+		);
+
+		await this._pc.setRemoteDescription(answer);
 	}
 
 	async replaceTrack(
 		localId: string,
 		track: MediaStreamTrack | null
 	): Promise<void> {
+		this.assertNotClosed();
 		this.assertSendDirection();
 
 		if (track) {
@@ -542,31 +673,20 @@ export class Chrome67 extends HandlerInterface {
 			logger.debug('replaceTrack() [localId:%s, no track]', localId);
 		}
 
-		const rtpSender = this._mapSendLocalIdRtpSender.get(localId);
+		const transceiver = this._mapMidTransceiver.get(localId);
 
-		if (!rtpSender) {
-			throw new Error('associated RTCRtpSender not found');
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
 		}
 
-		const oldTrack = rtpSender.track;
-
-		await rtpSender.replaceTrack(track);
-
-		// Remove the old track from the local stream.
-		if (oldTrack) {
-			this._sendStream.removeTrack(oldTrack);
-		}
-
-		// Add the new track to the local stream.
-		if (track) {
-			this._sendStream.addTrack(track);
-		}
+		await transceiver.sender.replaceTrack(track);
 	}
 
 	async setMaxSpatialLayer(
 		localId: string,
 		spatialLayer: number
 	): Promise<void> {
+		this.assertNotClosed();
 		this.assertSendDirection();
 
 		logger.debug(
@@ -575,13 +695,13 @@ export class Chrome67 extends HandlerInterface {
 			spatialLayer
 		);
 
-		const rtpSender = this._mapSendLocalIdRtpSender.get(localId);
+		const transceiver = this._mapMidTransceiver.get(localId);
 
-		if (!rtpSender) {
-			throw new Error('associated RTCRtpSender not found');
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
 		}
 
-		const parameters = rtpSender.getParameters();
+		const parameters = transceiver.sender.getParameters();
 
 		parameters.encodings.forEach(
 			(encoding: RTCRtpEncodingParameters, idx: number) => {
@@ -593,10 +713,31 @@ export class Chrome67 extends HandlerInterface {
 			}
 		);
 
-		await rtpSender.setParameters(parameters);
+		await transceiver.sender.setParameters(parameters);
+
+		this._remoteSdp!.muxMediaSectionSimulcast(localId, parameters.encodings);
+
+		const offer = await this._pc.createOffer();
+
+		logger.debug(
+			'setMaxSpatialLayer() | calling pc.setLocalDescription() [offer:%o]',
+			offer
+		);
+
+		await this._pc.setLocalDescription(offer);
+
+		const answer = { type: 'answer', sdp: this._remoteSdp!.getSdp() };
+
+		logger.debug(
+			'setMaxSpatialLayer() | calling pc.setRemoteDescription() [answer:%o]',
+			answer
+		);
+
+		await this._pc.setRemoteDescription(answer);
 	}
 
 	async setRtpEncodingParameters(localId: string, params: any): Promise<void> {
+		this.assertNotClosed();
 		this.assertSendDirection();
 
 		logger.debug(
@@ -605,13 +746,13 @@ export class Chrome67 extends HandlerInterface {
 			params
 		);
 
-		const rtpSender = this._mapSendLocalIdRtpSender.get(localId);
+		const transceiver = this._mapMidTransceiver.get(localId);
 
-		if (!rtpSender) {
-			throw new Error('associated RTCRtpSender not found');
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
 		}
 
-		const parameters = rtpSender.getParameters();
+		const parameters = transceiver.sender.getParameters();
 
 		parameters.encodings.forEach(
 			(encoding: RTCRtpEncodingParameters, idx: number) => {
@@ -619,19 +760,40 @@ export class Chrome67 extends HandlerInterface {
 			}
 		);
 
-		await rtpSender.setParameters(parameters);
+		await transceiver.sender.setParameters(parameters);
+
+		this._remoteSdp!.muxMediaSectionSimulcast(localId, parameters.encodings);
+
+		const offer = await this._pc.createOffer();
+
+		logger.debug(
+			'setRtpEncodingParameters() | calling pc.setLocalDescription() [offer:%o]',
+			offer
+		);
+
+		await this._pc.setLocalDescription(offer);
+
+		const answer = { type: 'answer', sdp: this._remoteSdp!.getSdp() };
+
+		logger.debug(
+			'setRtpEncodingParameters() | calling pc.setRemoteDescription() [answer:%o]',
+			answer
+		);
+
+		await this._pc.setRemoteDescription(answer);
 	}
 
 	async getSenderStats(localId: string): Promise<RTCStatsReport> {
+		this.assertNotClosed();
 		this.assertSendDirection();
 
-		const rtpSender = this._mapSendLocalIdRtpSender.get(localId);
+		const transceiver = this._mapMidTransceiver.get(localId);
 
-		if (!rtpSender) {
-			throw new Error('associated RTCRtpSender not found');
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
 		}
 
-		return rtpSender.getStats();
+		return transceiver.sender.getStats();
 	}
 
 	async sendDataChannel({
@@ -641,6 +803,7 @@ export class Chrome67 extends HandlerInterface {
 		label,
 		protocol,
 	}: HandlerSendDataChannelOptions): Promise<HandlerSendDataChannelResult> {
+		this.assertNotClosed();
 		this.assertSendDirection();
 
 		const options = {
@@ -648,7 +811,6 @@ export class Chrome67 extends HandlerInterface {
 			id: this._nextSendSctpStreamId,
 			ordered,
 			maxPacketLifeTime,
-			maxRetransmitTime: maxPacketLifeTime, // NOTE: Old spec.
 			maxRetransmits,
 			protocol,
 		};
@@ -711,19 +873,23 @@ export class Chrome67 extends HandlerInterface {
 	async receive(
 		optionsList: HandlerReceiveOptions[]
 	): Promise<HandlerReceiveResult[]> {
+		this.assertNotClosed();
 		this.assertRecvDirection();
 
 		const results: HandlerReceiveResult[] = [];
+		const mapLocalId: Map<string, string> = new Map();
 
 		for (const options of optionsList) {
 			const { trackId, kind, rtpParameters, streamId } = options;
 
 			logger.debug('receive() [trackId:%s, kind:%s]', trackId, kind);
 
-			const mid = kind;
+			const localId = rtpParameters.mid ?? String(this._mapMidTransceiver.size);
+
+			mapLocalId.set(trackId, localId);
 
 			this._remoteSdp!.receive({
-				mid,
+				mid: localId,
 				kind,
 				offerRtpParameters: rtpParameters,
 				streamId: streamId ?? rtpParameters.rtcp!.cname!,
@@ -740,14 +906,31 @@ export class Chrome67 extends HandlerInterface {
 
 		await this._pc.setRemoteDescription(offer);
 
+		for (const options of optionsList) {
+			const { trackId, onRtpReceiver } = options;
+
+			if (onRtpReceiver) {
+				const localId = mapLocalId.get(trackId);
+				const transceiver = this._pc
+					.getTransceivers()
+					.find((t: RTCRtpTransceiver) => t.mid === localId);
+
+				if (!transceiver) {
+					throw new Error('transceiver not found');
+				}
+
+				onRtpReceiver(transceiver.receiver);
+			}
+		}
+
 		let answer = await this._pc.createAnswer();
 		const localSdpObject = sdpTransform.parse(answer.sdp);
 
 		for (const options of optionsList) {
-			const { kind, rtpParameters } = options;
-			const mid = kind;
+			const { trackId, rtpParameters } = options;
+			const localId = mapLocalId.get(trackId);
 			const answerMediaObject = localSdpObject.media.find(
-				(m: any) => String(m.mid) === mid
+				(m: any) => String(m.mid) === localId
 			);
 
 			// May need to modify codec parameters in the answer based on codec
@@ -775,29 +958,24 @@ export class Chrome67 extends HandlerInterface {
 		await this._pc.setLocalDescription(answer);
 
 		for (const options of optionsList) {
-			const { kind, trackId, rtpParameters } = options;
-			const localId = trackId;
-			const mid = kind;
-			const rtpReceiver = this._pc
-				.getReceivers()
-				.find((r: RTCRtpReceiver) => r.track && r.track.id === localId);
+			const { trackId } = options;
+			const localId = mapLocalId.get(trackId)!;
+			const transceiver = this._pc
+				.getTransceivers()
+				.find((t: RTCRtpTransceiver) => t.mid === localId);
 
-			if (!rtpReceiver) {
-				throw new Error('new RTCRtpReceiver not');
+			if (!transceiver) {
+				throw new Error('new RTCRtpTransceiver not found');
+			} else {
+				// Store in the map.
+				this._mapMidTransceiver.set(localId, transceiver);
+
+				results.push({
+					localId,
+					track: transceiver.receiver.track,
+					rtpReceiver: transceiver.receiver,
+				});
 			}
-
-			// Insert into the map.
-			this._mapRecvLocalIdInfo.set(localId, {
-				mid,
-				rtpParameters,
-				rtpReceiver,
-			});
-
-			results.push({
-				localId,
-				track: rtpReceiver.track,
-				rtpReceiver,
-			});
 		}
 
 		return results;
@@ -806,19 +984,20 @@ export class Chrome67 extends HandlerInterface {
 	async stopReceiving(localIds: string[]): Promise<void> {
 		this.assertRecvDirection();
 
+		if (this._closed) {
+			return;
+		}
+
 		for (const localId of localIds) {
 			logger.debug('stopReceiving() [localId:%s]', localId);
 
-			const { mid, rtpParameters } =
-				this._mapRecvLocalIdInfo.get(localId) ?? {};
+			const transceiver = this._mapMidTransceiver.get(localId);
 
-			// Remove from the map.
-			this._mapRecvLocalIdInfo.delete(localId);
+			if (!transceiver) {
+				throw new Error('associated RTCRtpTransceiver not found');
+			}
 
-			this._remoteSdp!.planBStopReceiving({
-				mid: mid!,
-				offerRtpParameters: rtpParameters!,
-			});
+			this._remoteSdp!.closeMediaSection(transceiver.mid!);
 		}
 
 		const offer = { type: 'offer', sdp: this._remoteSdp!.getSdp() };
@@ -838,32 +1017,95 @@ export class Chrome67 extends HandlerInterface {
 		);
 
 		await this._pc.setLocalDescription(answer);
+
+		for (const localId of localIds) {
+			this._mapMidTransceiver.delete(localId);
+		}
 	}
 
-	async pauseReceiving(
-		// eslint-disable-next-line @typescript-eslint/no-unused-vars
-		localIds: string[]
-	): Promise<void> {
-		// Unimplemented.
+	async pauseReceiving(localIds: string[]): Promise<void> {
+		this.assertNotClosed();
+		this.assertRecvDirection();
+
+		for (const localId of localIds) {
+			logger.debug('pauseReceiving() [localId:%s]', localId);
+
+			const transceiver = this._mapMidTransceiver.get(localId);
+
+			if (!transceiver) {
+				throw new Error('associated RTCRtpTransceiver not found');
+			}
+
+			transceiver.direction = 'inactive';
+			this._remoteSdp!.pauseMediaSection(localId);
+		}
+
+		const offer = { type: 'offer', sdp: this._remoteSdp!.getSdp() };
+
+		logger.debug(
+			'pauseReceiving() | calling pc.setRemoteDescription() [offer:%o]',
+			offer
+		);
+
+		await this._pc.setRemoteDescription(offer);
+
+		const answer = await this._pc.createAnswer();
+
+		logger.debug(
+			'pauseReceiving() | calling pc.setLocalDescription() [answer:%o]',
+			answer
+		);
+
+		await this._pc.setLocalDescription(answer);
 	}
 
-	async resumeReceiving(
-		// eslint-disable-next-line @typescript-eslint/no-unused-vars
-		localIds: string[]
-	): Promise<void> {
-		// Unimplemented.
+	async resumeReceiving(localIds: string[]): Promise<void> {
+		this.assertNotClosed();
+		this.assertRecvDirection();
+
+		for (const localId of localIds) {
+			logger.debug('resumeReceiving() [localId:%s]', localId);
+
+			const transceiver = this._mapMidTransceiver.get(localId);
+
+			if (!transceiver) {
+				throw new Error('associated RTCRtpTransceiver not found');
+			}
+
+			transceiver.direction = 'recvonly';
+			this._remoteSdp!.resumeReceivingMediaSection(localId);
+		}
+
+		const offer = { type: 'offer', sdp: this._remoteSdp!.getSdp() };
+
+		logger.debug(
+			'resumeReceiving() | calling pc.setRemoteDescription() [offer:%o]',
+			offer
+		);
+
+		await this._pc.setRemoteDescription(offer);
+
+		const answer = await this._pc.createAnswer();
+
+		logger.debug(
+			'resumeReceiving() | calling pc.setLocalDescription() [answer:%o]',
+			answer
+		);
+
+		await this._pc.setLocalDescription(answer);
 	}
 
 	async getReceiverStats(localId: string): Promise<RTCStatsReport> {
+		this.assertNotClosed();
 		this.assertRecvDirection();
 
-		const { rtpReceiver } = this._mapRecvLocalIdInfo.get(localId) ?? {};
+		const transceiver = this._mapMidTransceiver.get(localId);
 
-		if (!rtpReceiver) {
-			throw new Error('associated RTCRtpReceiver not found');
+		if (!transceiver) {
+			throw new Error('associated RTCRtpTransceiver not found');
 		}
 
-		return rtpReceiver.getStats();
+		return transceiver.receiver.getStats();
 	}
 
 	async receiveDataChannel({
@@ -871,17 +1113,21 @@ export class Chrome67 extends HandlerInterface {
 		label,
 		protocol,
 	}: HandlerReceiveDataChannelOptions): Promise<HandlerReceiveDataChannelResult> {
+		this.assertNotClosed();
 		this.assertRecvDirection();
 
-		const { streamId, ordered, maxPacketLifeTime, maxRetransmits } =
-			sctpStreamParameters;
+		const {
+			streamId,
+			ordered,
+			maxPacketLifeTime,
+			maxRetransmits,
+		}: SctpStreamParameters = sctpStreamParameters;
 
 		const options = {
 			negotiated: true,
 			id: streamId,
 			ordered,
 			maxPacketLifeTime,
-			maxRetransmitTime: maxPacketLifeTime, // NOTE: Old spec.
 			maxRetransmits,
 			protocol,
 		};
@@ -893,7 +1139,7 @@ export class Chrome67 extends HandlerInterface {
 		// If this is the first DataChannel we need to create the SDP offer with
 		// m=application section.
 		if (!this._hasDataChannelMediaSection) {
-			this._remoteSdp!.receiveSctpAssociation({ oldDataChannelSpec: true });
+			this._remoteSdp!.receiveSctpAssociation();
 
 			const offer = { type: 'offer', sdp: this._remoteSdp!.getSdp() };
 
@@ -958,6 +1204,12 @@ export class Chrome67 extends HandlerInterface {
 		});
 
 		this._transportReady = true;
+	}
+
+	private assertNotClosed(): void {
+		if (this._closed) {
+			throw new InvalidStateError('method called in a closed handler');
+		}
 	}
 
 	private assertSendDirection(): void {
