+++ src/test/uaTestCases.ts
@@ -28,29 +28,9 @@ export const uaTestCases: UATestCase[] = [
 		expect: 'Chrome74',
 	},
 	{
-		desc: 'Microsoft Edge 44',
-		ua: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.140 Safari/537.36 Edge/18.17763',
-		expect: 'Edge11',
-	},
-	{
-		desc: 'Firefox 68 (Android)',
-		ua: 'Mozilla/5.0 (Android 10; Mobile; rv:68.10.0) Gecko/68.10.0 Firefox/68.10.0',
-		expect: 'Firefox60',
-	},
-	{
 		desc: 'In-app WebView (Android)',
 		ua: 'Mozilla/5.0 (Linux; Android 11; G91 Pro Build/RP1A.200720.011; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/114.0.5735.130 Mobile Safari/537.36',
 		expect: 'Chrome111',
-	},
-	{
-		desc: 'Safari 11',
-		ua: 'Mozilla/5.0 (Macintosh; Intel Mac OS X 11_13) AppleWebKit/537.36 (KHTML, like Gecko) Version/11.0.92 Safari/619.28',
-		expect: 'Safari11',
-	},
-	{
-		desc: 'Safari 11 (iPad)',
-		ua: 'Mozilla/5.0 (iPad; CPU OS 14_4 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/11.0 Mobile/15E148 Safari/604.1',
-		expect: 'Safari11',
 	},
 	{
 		desc: 'Brave',
@@ -61,11 +41,6 @@ export const uaTestCases: UATestCase[] = [
 		desc: 'In-app WebView (Android) (Facebook)',
 		ua: 'Mozilla/5.0 (Linux; Android 12; SM-S908U1 Build/SP1A.210812.016; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/100.0.4896.88 Mobile Safari/537.36 [FB_IAB/FB4A;FBAV/377.0.0.22.107;]',
 		expect: 'Chrome74',
-	},
-	{
-		desc: 'Firefox (Linux)',
-		ua: 'Mozilla/5.0 (X11; Ubuntu; Linux i686; rv:109.0) Gecko/20100101 Firefox/114.0',
-		expect: 'Firefox60',
 	},
 	{
 		desc: 'Firefox (iOS) - Unsupported',
